backend/requirements.txt
fastapi
httpx
orjson
faiss-cpu
numpy
pandas
langchain
langchain-core
langchain-community
underthesea
sentence_transformers